Adds support for having multiple interface state delegates.
Hopefully in a performant way.
